third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-summary-retainers-expected.txt

Tests retainers view.
    - Number of retainers of an A object must be 2 (A itself and B).
    - When an object has just one retainer it must be expanded automatically until
      there's an object having two or more retainers.
    - Test the expansion of a long retainment chain is limited by a certain level.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testRetainersView

Running: testRetainersAutoExpandSingleRetainer

Running: testRetainersAutoExpandSingleRetainerLimit

Profiler was disabled.
